+++ conan/api/conan_api.py
@@ -17,15 +17,15 @@ from conan.api.subapi.remotes import RemotesAPI
 from conan.api.subapi.remove import RemoveAPI
 from conan.api.subapi.search import SearchAPI
 from conan.api.subapi.upload import UploadAPI
-from conans.client.conf.required_version import check_required_conan_version
 from conans.client.migrations import ClientMigrator
 from conans.client.userio import init_colorama
 from conans.errors import ConanException
 from conans.model.version import Version
 from conan.internal.paths import get_conan_user_home
+from conans.model.version_range import validate_conan_version
 
 
-class ConanAPI(object):
+class ConanAPI:
     def __init__(self, cache_folder=None):
 
         version = sys.version_info
@@ -59,4 +59,6 @@ class ConanAPI(object):
         self.lockfile = LockfileAPI(self)
         self.local = LocalAPI(self)
 
-        check_required_conan_version(self.config.global_conf)
+        required_range_new = self.config.global_conf.get("core:required_conan_version")
+        if required_range_new:
+            validate_conan_version(required_range_new)
