+++ conan/internal/api/detect/detect_vs.py
