+++ conan/internal/api/profile/detect.py
@@ -1,5 +1,5 @@
 from conan.api.output import ConanOutput
-from conan.internal.api.detect_api import detect_os, detect_arch, default_msvc_runtime, \
+from conan.internal.api.detect.detect_api import detect_os, detect_arch, default_msvc_runtime, \
     detect_libcxx, detect_cppstd, detect_default_compiler, default_compiler_version
