+++ test/unittests/client/conf/detect/test_gcc_compiler.py
@@ -4,7 +4,7 @@ import mock
 from parameterized import parameterized
 
 
-from conan.internal.api.detect_api import detect_gcc_compiler
+from conan.internal.api.detect.detect_api import detect_gcc_compiler
 
 
 class GCCCompilerTestCase(unittest.TestCase):
@@ -12,7 +12,7 @@ class GCCCompilerTestCase(unittest.TestCase):
     @parameterized.expand([("10",), ("4.2",), ('7', )])
     def test_detect_gcc_10(self, version):
         with mock.patch("platform.system", return_value="Linux"):
-            with mock.patch("conan.internal.api.detect_api.detect_runner", return_value=(0, version)):
+            with mock.patch("conan.internal.api.detect.detect_api.detect_runner", return_value=(0, version)):
                 compiler, installed_version, compiler_exe = detect_gcc_compiler()
         self.assertEqual(compiler, 'gcc')
         self.assertEqual(installed_version, version)
