+++ test/unittests/tools/cmake/test_cmake_install.py
@@ -1,6 +1,6 @@
+from conan.internal.default_settings import default_settings_yml
 from conan.tools.cmake import CMake
 from conan.tools.cmake.presets import write_cmake_presets
-from conans.client.conf import get_default_settings_yml
 from conans.model.conf import Conf
 from conans.model.settings import Settings
 from conan.test.utils.mocks import ConanFileMock
@@ -13,7 +13,7 @@ def test_run_install_component():
     Issue related: https://github.com/conan-io/conan/issues/10359
     """
     # Load some generic windows settings
-    settings = Settings.loads(get_default_settings_yml())
+    settings = Settings.loads(default_settings_yml)
     settings.os = "Windows"
     settings.arch = "x86"
     settings.build_type = "Release"
@@ -42,7 +42,7 @@ def test_run_install_strip():
     Issue related: https://github.com/conan-io/conan/issues/14166
     """
 
-    settings = Settings.loads(get_default_settings_yml())
+    settings = Settings.loads(default_settings_yml)
     settings.os = "Linux"
     settings.arch = "x86_64"
     settings.build_type = "Release"
@@ -71,7 +71,7 @@ def test_run_install_cli_args():
     Issue related: https://github.com/conan-io/conan/issues/14235
     """
 
-    settings = Settings.loads(get_default_settings_yml())
+    settings = Settings.loads(default_settings_yml)
     settings.os = "Linux"
     settings.arch = "x86_64"
     settings.build_type = "Release"
@@ -99,7 +99,7 @@ def test_run_install_cli_args_strip():
     Issue related: https://github.com/conan-io/conan/issues/14235
     """
 
-    settings = Settings.loads(get_default_settings_yml())
+    settings = Settings.loads(default_settings_yml)
     settings.os = "Linux"
     settings.arch = "x86_64"
     settings.build_type = "Release"
