+++ test/unittests/util/detected_architecture_test.py
@@ -3,7 +3,7 @@ import unittest
 
 from parameterized import parameterized
 
-from conan.internal.api.detect_api import detect_arch
+from conan.internal.api.detect.detect_api import detect_arch
 
 
 class DetectedArchitectureTest(unittest.TestCase):
@@ -41,14 +41,14 @@ class DetectedArchitectureTest(unittest.TestCase):
         with mock.patch("platform.machine", mock.MagicMock(return_value='00FB91F44C00')),\
                 mock.patch("platform.processor", mock.MagicMock(return_value='powerpc')),\
                 mock.patch("platform.system", mock.MagicMock(return_value='AIX')),\
-                mock.patch("conan.internal.api.detect_api._get_aix_conf", mock.MagicMock(return_value='32')),\
+                mock.patch("conan.internal.api.detect.detect_api._get_aix_conf", mock.MagicMock(return_value='32')),\
                 mock.patch('subprocess.check_output', mock.MagicMock(return_value='7.1.0.0')):
             self.assertEqual('ppc32', detect_arch())
 
         with mock.patch("platform.machine", mock.MagicMock(return_value='00FB91F44C00')),\
                 mock.patch("platform.processor", mock.MagicMock(return_value='powerpc')),\
                 mock.patch("platform.system", mock.MagicMock(return_value='AIX')),\
-                mock.patch("conan.internal.api.detect_api._get_aix_conf", mock.MagicMock(return_value='64')),\
+                mock.patch("conan.internal.api.detect.detect_api._get_aix_conf", mock.MagicMock(return_value='64')),\
                 mock.patch('subprocess.check_output', mock.MagicMock(return_value='7.1.0.0')):
             self.assertEqual('ppc64', detect_arch())
